refactor(id3v2): extract shared container ID3v2 encoder, drop dead validate exports
The MP4/AIFF/AAC encoders each repeated the same ID3v2 options-building,
encodingIndex, validate, and encode sequence, differing only in default
padding/unsynch. Extract it into utils/container.mjs#encodeID3v2.

The per-module validate() exports were never imported anywhere (each
encoder calls ID3v2.validate directly) — remove them as dead code.
